ALS-9090: Add old and new integration test coverage, initial implementation (#156)

+++ services/pic-sure-hpds/client-api/src/main/java/edu/harvard/hms/dbmi/avillach/hpds/data/query/v3/PhenotypicFilter.java
@@ -2,5 +2,5 @@ package edu.harvard.hms.dbmi.avillach.hpds.data.query.v3;
 
 import java.util.List;
 
-public record PhenotypicFilter(PhenotypicFilterType phenotypicFilterType, String conceptPath, List<String> values, Float min, Float max, Boolean not) implements PhenotypicClause {
+public record PhenotypicFilter(PhenotypicFilterType phenotypicFilterType, String conceptPath, List<String> values, Double min, Double max, Boolean not) implements PhenotypicClause {
 }
